♻️ Refactor import_source
